Fixed disk emulation to handle quarter track seeking.
It's not very accurate as it's set to zap the head instantly to where
the phase solenoids indicate, as opposed to moving from the start
position to the destination position in a finite amount of time (which
wouldn't be all that difficult to emulate).  But most things tend to be
pretty well behaved and so this approach seems to work OK.

Also added preliminary hard disk emulation.